feat: [Novo] Convert more routes
