+++ src/v2/Apps/getAppNovoRoutes.tsx
@@ -4,10 +4,12 @@ import { artistRoutes } from "v2/Apps/Artist/artistRoutes"
 import { artistSeriesRoutes } from "./ArtistSeries/artistSeriesRoutes"
 import { artistsRoutes } from "v2/Apps/Artists/artistsRoutes"
 import { artworkRoutes } from "v2/Apps/Artwork/artworkRoutes"
+import { auctionsRoutes } from "./Auctions/auctionsRoutes"
 import { collectRoutes } from "v2/Apps/Collect/collectRoutes"
 import { consignRoutes } from "v2/Apps/Consign/consignRoutes"
 import { conversationRoutes } from "v2/Apps/Conversation/conversationRoutes"
 import { debugRoutes } from "./Debug/debugRoutes"
+import { exampleRoutes } from "./Example/exampleRoutes"
 import { fairRoutes } from "v2/Apps/Fair/fairRoutes"
 import { fairsRoutes } from "v2/Apps/Fairs/fairsRoutes"
 import { featureRoutes } from "v2/Apps/Feature/featureRoutes"
@@ -38,14 +40,23 @@ export function getAppNovoRoutes(): RouteConfig[] {
       },
       {
         converted: true,
+        routes: auctionsRoutes,
+      },
+      {
+        converted: true,
         routes: collectRoutes,
       },
       {
+        converted: true,
         routes: consignRoutes,
       },
       {
         converted: true,
         routes: conversationRoutes,
+      },
+      {
+        converted: true,
+        routes: exampleRoutes,
       },
       {
         converted: true,
@@ -69,6 +80,7 @@ export function getAppNovoRoutes(): RouteConfig[] {
         routes: purchaseRoutes,
       },
       {
+        converted: true,
         routes: searchRoutes,
       },
       {
