+++ src/v2/Apps/getAppRoutes.tsx
@@ -5,20 +5,20 @@ import { RouteConfig } from "found"
 // import { artistsRoutes } from "v2/Apps/Artists/artistsRoutes"
 import { artworkRoutes } from "v2/Apps/Artwork/artworkRoutes"
 // import { collectRoutes } from "v2/Apps/Collect/collectRoutes"
-import { consignRoutes } from "v2/Apps/Consign/consignRoutes"
+// import { consignRoutes } from "v2/Apps/Consign/consignRoutes"
 // import { conversationRoutes } from "v2/Apps/Conversation/conversationRoutes"
 import { debugRoutes } from "./Debug/debugRoutes"
-import { exampleRoutes } from "./Example/exampleRoutes"
+// import { exampleRoutes } from "./Example/exampleRoutes"
 // import { fairRoutes } from "v2/Apps/Fair/fairRoutes"
 // import { fairsRoutes } from "v2/Apps/Fairs/fairsRoutes"
 // import { featureRoutes } from "v2/Apps/Feature/featureRoutes"
 import { identityVerificationRoutes } from "v2/Apps/IdentityVerification/identityVerificationRoutes"
 import { orderRoutes } from "v2/Apps/Order/orderRoutes"
 import { purchaseRoutes } from "v2/Apps/Purchase/purchaseRoutes"
-import { searchRoutes } from "v2/Apps/Search/searchRoutes"
+// import { searchRoutes } from "v2/Apps/Search/searchRoutes"
 // import { showRoutes } from "v2/Apps/Show/showRoutes"
 // import { viewingRoomRoutes } from "./ViewingRoom/viewingRoomRoutes"
-import { auctionsRoutes } from "./Auctions/auctionsRoutes"
+// import { auctionsRoutes } from "./Auctions/auctionsRoutes"
 
 export function getAppRoutes(): RouteConfig[] {
   return buildAppRoutes([
@@ -37,23 +37,23 @@ export function getAppRoutes(): RouteConfig[] {
     {
       routes: artworkRoutes,
     },
-    {
-      routes: auctionsRoutes,
-    },
     // NOTE: Converted to use NOVO template.
+    // {
+    //   routes: auctionsRoutes,
+    // },
     // {
     //   routes: collectRoutes,
     // },
-    {
-      routes: consignRoutes,
-    },
+    // {
+    //   routes: consignRoutes,
+    // },
     // NOTE: Converted to use NOVO template.
     // {
     //   routes: conversationRoutes,
     // },
-    {
-      routes: exampleRoutes,
-    },
+    // {
+    //   routes: exampleRoutes,
+    // },
     // NOTE: Converted to use NOVO template.
     // {
     //   routes: fairRoutes,
@@ -73,10 +73,10 @@ export function getAppRoutes(): RouteConfig[] {
     {
       routes: purchaseRoutes,
     },
-    {
-      routes: searchRoutes,
-    },
     // NOTE: Converted to use NOVO template.
+    // {
+    //   routes: searchRoutes,
+    // },
     // {
     //   routes: showRoutes,
     // },
